+++ app/src/main/java/kz/narxoz/canvasdiplom/models/Prerequisite.kt
@@ -4,8 +4,8 @@ import java.math.BigInteger
 import java.sql.Timestamp
 
 data class Prerequisite(
-    val courseId: BigInteger,
-    val prerequisiteCourseId: BigInteger,
+    val courseId: Long,
+    val prerequisiteCourseId: Long,
 
     val createdAt: Timestamp,
     val updatedAt: Timestamp
